readme: Fix some bugs/omissions in the README.
When setting up prost-wkt, I found one bug and one omission in the
README examples.

The bug is that if you use the `#[serde(default)]` macro across
everything, any enums that you have won't compile because they don't
support this option.

The omission was that you need to have an explicit dependency on `serde`
with the derive feature enabled.